feat(cmd-api-server): fully dynamic plugin imports via config file/env/CLI
It leverages the PluginRegistry to its full potential. Had to slightly refactor how the plugin
factories are exported from the plugin packages but it is a non-breaking change.

Also: decoupled
the consortium web service plugins tests from the plugin itself (e.g. moved it to a separate
package) so that circular package dependencies are avoided where with cmd-api-server

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>