Chore/housekeeping prettier ignore ci and build fixes (#148)
* docs(template-launch-json): cmd-api-server launcher task updated after the BIF -> Cactus rename

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>

* style: fixing typos in spec files, readmes, etc

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>

* fix(cmd-api-server): schema for the plugin array custom type

The loosely typed "options" property had validation issues when passed in through a config file.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>

* build(plugin-web-service-consortium): fix the order of tasks by having the SDK generated in pretsc

This way the generated SDK code gets reflected in the tsc compilation process

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>

* ci(travis): include NodeJS 12 and 14 with 10 among the tested versions

perf(ci-sh): uses npm ci instead of npm install which is designed for CI environments and therefore
is slightly faster.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>

* build(webpack,tsc): add webpack cache-loader and tsc incremental flag to speed builds up

caching is no magic bullet so we will have to monitor how this affects developer flow on the long
run and if theres any annoying side effects with cached builds hanging around longer than they
should be (e.g. cache invalidation fail)

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>

* build(package-json): refactor how tests are executed to get a unified view of test coverage

Before this, tests were executed on a per package level and dumped their coverage report somewhere
in the middle of the process, individually. After this you get a unified report at the very end
showing the total coverage across all packages.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>

* chore(prettierignore): make prettier ignore the new contribs folder so (do not format it)

I missed this when the original PR introducing the contribs folder was accepted.
We do not want to
work on the original codebases anymore
so having prettier leave it as is is probably the
best/easiest.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>

* build(tslint): cockpit tslint will not enforce double quotes to match prettier defaults

Previously it was single quotes and the auto formatter was stuck trying to please both (single vs
double). Does not make much difference to me which one we use but some sort of uniformity is needed
for sure.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>