feat(common): add utility class Strings with replaceAll
This is a method that is used pretty often but not included
in the Javascript language itself.
What the method does is what you would expect, it finds
all occurrences of a substring and replaces all of them with
another one. The deafult Javascript String.replace function
only does this with the first occurrence by default so it is not
a good solution.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>