build(webpack): adds shebang loader
Added a shebang-loader (webpack plugin) to avoid
webpack disappearing the hashbang from the standard bin
script of the example package.
Without this, if you have the shebang at the beginning
of your Typescript file such as this
And then build the project, it would just disappear, but
with the shebang loader it is preserved.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>