refactor: 💡 In-memory keychain getEncryptionAlgorithm() => null
Previously this method would return a string expressing in English prose
that it does not support encryption and therefore there are no
algorithms there either. It is deemed safer to return null because that
way no one could construe that there is a valid value in there (which
could indeed happen with the previous version if one was just checking
whether a truthy value has been returned by the method or not).

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>