docs(examples): supply chain app example
The idea is to have a very quickly testable demonstration
of how Cactus can be used to develop an (overly simplistic)
business use case that involves transacting over multiple
ledgers. This example does not yet provide an exhaustive
example on how to implement atomic swaps for example,
it is purely aimed at showing how code can be written
with the help of Cactus to perform basic tasks on
different ledgers.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>