ci(github): migrate to Github Actions for CI
Changes:
1. Deletes the .travis.yml file meaning that
Travis will no longer be our CI runner
2. Adds the .github/workflows/ci.yml file
meaning that Github Actions is stepping in as
the new Travis from now on.
3. Starts skipping a couple of tests which have
issues with passing temporarily due to pending
pull requests in the review queue.

Docker is not supported on macOS and there is
no end in sight to this problem either:
https://github.community/t/why-is-docker-not-installed-on-macos/17017

Windows has a path syntax issue in the npm script
responsible for generating the API clients so
for now that is disabled as well as macOS. We expect
to be able to solve this in the near future though
(unlike what is happening with macOS...)

Resolves #419

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>