fix(test-tooling): fabric AIO image docker in docker support
Fixes #279

Note: Although this image now uses Docker in Docker (DinD)
it does not actually solve the problem of randomized
ports completely because as it turns out there's a second
issue with the randomized ports namely that Fabric's
service discovery doesn't have a port mapping feature in it
so we are not able to specify the associations between
the public and private ports where the public (host) ports
are randomized and the private ones are the ones
returned by the service discovery algorithms.
For example, if you start the new AIO image with randomized
ports, then it will run one of the peers on port 7051 of the
AIO container and the host will map that to something random
typically somewhere in the 30000 to 40000 port range.
When the Fabric connector (that's running on the host machine)
instantiates a Fabric Gateway object of the Fabric Node SDK
it performs service discovery where the peer is described
by the service discovery as listening on port 7051, but from
the network of the host (where the Fabric connector is)
that port is not correct because the real one is the random
one as explained in the above paragraph.
What we need to solve this is a way to inject into the service
discovery mechanism our own port mappings so that the
Fabric connector related tests can run in parallel and the
CI can stop being flaky.

Because of what is explained above, in this commit we also
skip the Fabric AIO image dependent test cases until they
can be fixed via refactoring the test cases and the test
ledger class in a way so that it works with the new AIO
image. These changes will need to have some method deletions
and renames which will make the change quite big on its own
and therefore will go in a separate commit which will also
re-activate the test cases that are currently being skipped
as per this current change.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>