feat(core-api): getConsensusAlgorithmFamily() on connector API
The method is called getConsensusAlgorithmFamily()
because we don't need (for now) the level of granularity
that is provided by having a method that points to the exact
algorithm used by the ledger (can be added later).
Why? Because for now we are mostly interested in whether an
algorithm family guarantees transaction finality or not and this
can be determined just from the family since for example it
does not make much of a difference if you are talking about
Bitcon's or Ethereum 1's proof of work, they both just do not
guarantee transaction finality the same way for all our intents
and purposes at present.

The added benefit of only dealing in families instead of specific
algorithms is that we can hardcode the answers instead of
having to query the ledger or rely on operators to provide
this information via configuration (again, we'll probably end
up needing this in the future anyway, but for the upcoming
work the families should be enough).

Fixes #355

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>