feat(common): servers.startOnPort() host arg #527
Extends the existing functionality of the utility method so that it can
accept a 'host' parameter where callers can
specify if they want to bind to the wildcard
network interface of `0.0.0.0` for example.

The default option of the host argument is
localhost so that we are *secure by default*
even if this method is only used for testing
code, it cannot hurt to be on the safe side.

Fixes #527

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>