fix(ci): disk full issues on GitHub Action Workflow runner #698
Updates the test cases that use ledger containers to clean up
after themselves not just by stopping and deleting the containers
but by also doing a docker prune of the containers, images, networks
and volumes as well.
What we hope to accomplish here is that the GHA CI will stop
failing with disk full errors.

Fixes #698

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>