feat(tools): add standalone shell CLI tool to corda AIO image
The corda AIO docker image now downloads
the CLI tools jar for the standalone shell at
build time.
By default it downloads the 4.5 version of the
jar but this can be altered to any other version
through a dedicated build argument of the
container image definition.

The image built  from this commit is also tagged
(on top of the automatic tagging script) as
hyperledger/cactus-corda-4-6-all-in-one-obligation:2021-03-14-standalone-corda-shell

See more details here for what a standalone
shell is and how it can be used to execute
commands (some which are not available on
the RPC proxy so this tool is a must have for
certain operations such as gracefully stopping
nodes)
https://docs.corda.net/docs/corda-os/4.5/shell.html#standalone-shell

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>