docs(readme): add reference to BUILD.md
BUILD.md: add GIF showcasing the watch script

We've received feedback from multiple people that long build
times are tedious and reduce general productivity which is
(probably) hindering the growth of the community so here is
a video guide (GIF) on how fast builds can be achieved by
only rebuilding what you actually changed on the package level.

It also comes with support for regenerating the code that we
don't write ourselves but use the OpenAPI generator for it instead.
All the openapi.json files are being watched by the `npm run watch`
script for code changes the same way that Typescript files are being
watched so this just works out of the box.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>