fix(tools): use latest fabric bootstrap.sh script for AIO image
Fabric appears to have embraced inclusive naming for their branches and
deleted the previous default branch of theirs which all the older revisions
of their bootstrap.sh script is pointing to by way of attempint to check that
branch out.
This broke our containers at runtime because the old bootstrap scripts are
now crashing due to the non-existent branch.
Updated both the 1.4.x and the 2.x dockerfiles to use the current latest
bootstrap.sh script instead in order to alleviate the problem.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>