build: migrate to Yarn v1 from npm v7
Why?
It was reported by other maintainers that yarn behaves much
better in corporate proxy/firewalled environments.

Future plans: Once Berry (Yarn v2) is stable, we should take
a look at migrating onto that one. I did give it a try this
time to see if their release candidate happened to be stable
enough but it was not unfortunately (there were dependency
resolution issues that lead the Yarn v2 not installing
certain runtime dependencies at all so v1 it is for now)

Also:
1. Made the CI not fail fast along the test matrix dimensions
2. Sequentialized the test:all script to always run one test at a time
hoping to get some additional insight into the flaky tests this way.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>