fix(plugin-ledger-connector-quorum): no keychain endpoints
1. Endpoint invoke-contract without keychain was failing because,
on deploy, network information was saved inside the contractJson
object, which was keeping it on memory and for this reason it was
not failing calling the connector.

2. Now, when no keychain is used, instead of save the contracts
addresses in memory indexed by contractName, this parameter
should be sent in the request to avoid that there can only be a
single instance for each contract name. When keychain is used,
contract address is still saved indexed by contractName.

3. Openapi specs were changed to consider these changes,
including remove fields that are inside the contractJSON object
and definition of this object structure.

4. Includes openapi validation test for new endpoints.

Signed-off-by: Elena Izaguirre <e.izaguirre.equiza@accenture.com>