build: reset:node-modules script dry-run parameter dropped
Epic facepalm once again, congratulations to Peter!

The reset:node-modules script had one job, which was to
delete all node_modules directories in the entire repo.
It was on the right track, but the --dry-run flag was forgotten
in among the CLI arguments so this whole time the script was
not working *at all* because of it. Just brilliant... :-)

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>