test: supply-chain-backend-api-calls fails without config file #1516
Made it so that the test uses a config that does not specify a config
file path to be parsed by the config service so that it does not depend
on one existing on the file system when the test gets executed.
It works fine without the .config.json file because we anyway configure
everything via code within the test.

This test started failing while I was working on the Jest/Tap co-existence
pull request so it might have something to do with the migration, but it
does not look like it.

Fixes #1516

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>