ci(tools): fix iroha AIO image build flake on CI
Related to #1566 but does not actually fix it.
For now I just commented out the build step
of the flaky container image until we figure out
how to remedy the situation properly because
in the meantime I do not want to waste time
having to re-run the CI all the time because of the
flake.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>