docs(examples): explain supply chain live development
We should have documentation for how to develop the
supply chain app front-end with live-reloading enabled.
The ionic serve command has a built-in reverse-proxy
feature that makes this possible quite easily with
the right configuration in place.

How to achieve this is explained in the README.md
file of the supply chain app and the necessary updates
were made so that it actually works.

Fixes #2010

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>