docs(readme): cactus & weaver => cacti - new README and illustrations
Sketched out a tentative roadmap for the deeper merging of the
Cactus and Weaver code bases.

We plan on adding more changes to the readme soon after this has been
merged, but for now this is the initial "much better than nothing" draft
which we hope avoids confusion among prospective users who come in fresh
without having the contextual information/background story.

Also added a list of files to .dcilintignore because of some Corda test
network files that are generated and therefore cannot be altered to
comply with DCI-Lint

Signed-off-by: VRamakrishna <vramakr2@in.ibm.com>
Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>