chore: adding custom gitguardian check
related to #720

This includes gitguardian workflow + configuration file

This change was necessary even if we have stock git guardian workflow
action because it was not possible to be configured for exclusions.

The scanner was not getting triggered by pull_request_target as
expected to access the secrets environment variable so it was
returned to the original pull_request. By default, secrets is not
possible to be accessed by a pull request from a fork repository
unless it is merged to the main repository. Hence, the original
pull_request will work after merging.

The test tokens used in testing were all deleted to prevent further
usage of it.

Signed-off-by: ruzell22 <ruzell.vince.aquino@accenture.com>
Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>