feat(cacti-cmd-gui-app): add GUI to visualize Fabric, Ethereum blocks
Right now supporting two ledgers
- Hyperledger Fabric
- Ethereum

Updates from Peter to fix the CI checks:
1. The build:dev:backend script was broken, fixed it by removing the
clean, lint, codegen scripts which were removed earlier but this change
was trying to add them back for some reason.
2. Frontend package build (build:dev:frontend) was broken, fixed it by
correcting the polyfills for both the carbon accounting and the supply
chain example packages.
3. The build:dev:frontend script in the package.json in the root was
spawning more processes than it needed to. A single lerna invocation
with the appropriate filters can capture and run the build scripts in
all 3 of the packages. The entire point of lerna is to do this instead of
repeating 3 commands when you want to run 3 scripts in 3 packages so now
the way the npm script is written adheres to this convention.

Co-authored-by: Peter Somogyvari <peter.somogyvari@accenture.com>

Signed-off-by: Barnaba Pawelczak <barnaba.pawelczak@fujitsu.com>
Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>