build(plugin-keychain-vault): fix gradle error Redeclaration: ResponseType
Primary change:
---------------

1. The codegen script now runs the clean script prior to its own
execution automatically (so that we don't forget to run it).
If we don't run the clean script prior to the code generation, then
the code generation doesn't delete old files just dumps the new ones
next to the old ones which leads to problems after upgrading the generator:
A newer version of the generator usually ships with different templates
that have restructured their code files so we end up with duplicates
breaking the build (unless we do a clean prior to codegen).

As for the actual build issue that this task was about:
It was a cross-cutting concern because the root cause of it was that
we were not cleaning up properly prior to the codegen task but this
applied to all the 2 or so dozen kotlin projects that we generate.

So as part of the solution I've ran the clean script and ran codegen
again which now produces code that builds without issues.

Secondary changes:
-----------------

Modified the clean script in the root project the following ways:
1. It is now cleaning up generated kotlin OpenAPI code as well
(previously it was not doing this at all despite what the name of the
script might suggest...)
2. It excludes all the openapi-generator-ignore files from being deleted
which has the beneficial side-effect that they do not get re-generated
with their default contents which would be missing files that we want
to ignore in addition to them.
3. The clean script NO LONGER DELETES the
`src/main/typescript/generated/proto/protoc-gen-ts` sub-folder of the
cmd-api-server package because the contents of it were not being re-
generated by the `codegen` script for some reason. This is something
to be fixed properly in the future but for now I just had to cut my
losses and make it work.

This depends on the PR being merged above because they both touch on
the generated code but in different ways. If we merged this one first,
the other one would have its diff emptied out (stealing the show).

[skip ci]

Fixes #2730

Depends on https://github.com/hyperledger/cacti/pull/2729

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>