ci(github): add debugging to broken conditionals asset-exchange-besu-local
I thought I fixed this with
https://github.com/hyperledger/cacti/commit/11cbcb698274ac0f95cd096f5a9165acc500a3c6
but it turns out it just made it less broke. Now it is always running
all the matrix steps regardless of the code changes detected or not.

To help with debugging this further I've added some logging to the job
and planning on figuring out some fix based on that in future PRs.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>