tools: fix scripts - use run-time-error instead of run-time-error-cjs
1. The `-cjs` version of the package was put into place because in the
production packages we needed it to be able to import the dependency
despite us not using ESM just yet.
2. We did a blanket search and replace at the time which seemed like a
good idea but turns out it wasn't because there are some tooling scripts
(the ones in this diff) that are already using ESM and for those the
-cjs flavored package started breaking the code.
3. So this change is just a partial/selective/surgical revert of the
earlier project-wide migratin from run-time-error to run-time-error-cjs.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>