docs(examples/cbdc): migrate backend to Fabric v2.5.6 all-in-one image
1. The older versions of the AIO image are rusty (flaky to boot) and so
to increase the stability of the example application I considered it
prudent to upgrade to the latest and greatest (non-breaking) all-in-one
image that we have.
2. There might be other branches out there in development where the same
change is already pending, if this is the case then apologies, but I just
had to get this done right away because it had a direct dependency from
another pull request where I was upgrading web3 packages and I had to
have the example application up and running in order to verify that other
pull requests not messing anything up, so here we are.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>