build(connector-stellar): add skeleton for future Stellar connector
1. The diff is already quite large so I've decided to start with a smaller
pull request that is easier to review. Basic functionality is missing from
the connector because of the latter.
2. The build is confirmed to be working, so is code generation and tests.
3. We don't yet have a `StellarTestLedger` class in the test-utils package
and so we can't yet implement end to end testing but once that gets added
we'll be ready to go.
4. For now the public-api.ts and the index.ts and index.web.ts files are
not exporting anything so that we adhere to semantic versioning principles
in that we won't be doing any breaking changes with the sub-sequent pull
requests since there was nothing exported and nothing is unbreakable.
5. For now the package is set to private: true and the "access" within
`publishConfig` is set to restricted.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>