ci(github): yarn lint job now fails if formatting was needed
If someone sends a pull request that isn't formatted according to the
tools we have in place for auto-format code (ESLint, prettier) then after
this change the CI will fail and this way we don't end up in a situation
where running `yarn lint` on the main branch produces a diff of auto-formatted
code files (which is currently sadly the case).

At the same time as introducing this constraint I'm also applying the automatic
formatting on the code project wide so that the check can actually pass as
intended when this change is getting merged.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>