build(connector-corda): upgrade to Spring Boot v3.3.1
IMPORTANT: The project now needs JDK 17 and Gradle 8 for development.

1. The bean validation has been replaced to the jakarta flavor as mandated
by the spring boot upgrade.
2. Request bodies in the OpenAPI spec for Corda v4 were made mandatory
where applicable (which is most of the endpoints)
3. The Open API spec based request validation test case has been moved to
the new Corda v4.12 test case which now builds both the connector JVM app
and the AIO image at runtime so that we can guarantee that the latest
code changes are being verified.
4. Added quicker and easier steps in the readme of the corda connector
to run trivy scans on the .jar files instead of the container images.
The .jar files are 20x faster to build and scanning them instead of the
container images doesn't suffer from the problem that the dev build
dependencies are showing up in the scans (creating false positives)
5. Updated the CI to use the .jar file for scanning as well instead of
the container image.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>