docs(examples/supply-chain-app): fix resource cleanup of shutdown logic
1. Now you can run the supply chain app example locally and then hit CTRL+C
on the terminal and it will gracefully shut down all the containers hosting
the infrastructure and only after that the NodeJS process itself will exit.
2. Previously we had a bug where it wouldn't wait for the containers to wind
down and it left them running which was causing problems when people didn't
notice this behavior and their machines would get into this broken state where
the previous execution's containers were hanging around and one of them blocking
ports too.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>