feat(connector-corda): add vaultQueryV1 REST API operation + endpoint
Primary change(s):
-----------------

1. Adds a (for now) simplistic endpoint to query ledger state which allows
us to retrieve identifiers needed to construct flow requests in certain
use-cases.
2. This expands the number of supported use-cases for the Corda v4 (JVM)
connector.

Notes:
-------
Automated testing is work in progress, but it depends on a few other
big(er) changes so in an effort to limit the size of this pull request
I'm only submitting the steps for manual testing as documented in the
package's README.md file. These are steps which I have executed myself
and can vouch that they work as expected.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>