test(supply-chain-app-backend): fix via-npm-script.test.ts hanging
1. The test case was hanging at the end because it was waiting  for a
magic string to appear in the logs to indicate that the booting has finished.
2. The problems began when we changed the supply chain app's log message
from mentioning Cactus to Cacti which then invalidated the condition forever.
3. This left it hanging indefinitely and the test case was broken.
4. Now to fix the problem and avoid this happening again in the future
the supply chain app exports the log message pattern as a variable and the
test case imports that directly so if we change the log message in the
future it will automatically make sure that the test is also waiting for
the updated log message pattern to show up in the logs making this class
of bugs impossible to happen.
5. Also sneaking in a hot-fix for a missing test coverage environment
variable for the test package of the manual consortium plugin which was
uncovered by this pull request's execution by chance.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>