feat(consortium-static): new consortium plugin
* New plugin consortium-static, based on consortium-manual.
Long story short, the plugin allows for the addition of Cacti Nodes to
consortium at runtime. New node entities must belong to one of the
consortium entities (organizations) that were specified on consortium
creation. So it allows to add new nodes 'CactusNode', but not new
'ConsortiumMember' entities.

* New feature: add CactusNode to consortium at runtime
The code is based on the plugin-consortium-manual, with the new features
built on top of it. The process of adding a new node is conducted as
follows:
   1. The new node forges a request with information about itself and
   sends it to any Node in the consortium
   2. The receiver verifies the request, and if it is approved,
   broadcasts it to the rest of the consortium Nodes.
   3. Each node verifies the same request, and adds the new node to the
   consortium database. At the same time, the receiver node from point 2
   sends the consortium data to the new node, who updates it's database.

* Includes a new policy model
The package includes a policy model, which was developed based on the
ideas of the Policy Core Information Model
[RFC3060](https://www.rfc-editor.org/rfc/rfc3060) . The idea was to
introduce the idea of a multi-purpose policy framework that can be
leveraged by other packages in cacti.

Signed-off-by: eduv09 <eduardovasques10@tecnico.ulisboa.pt>