feat(common): add BigInt JSON replacer (number -> string)
This can be used by code on the back-end to be able to safely send
BigInt values through the  various transports (HTTP, gRPC, CRPC, etc.)

By default JSON.stringify will convert BigInt values to numbers which
can cause data loss if the number is too large. By using this replacer
function, we can ensure that BigInt values are always sent as strings.

It adds overhead on the client side to have to manually parse the BigInt
values, but it is a necessary trade-off for preserving data integrity.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>